vosk-stt/requirements.txt
# Vosk STT Dependencies for cPanel Python 3.13
# Install via: pip install -r requirements.txt

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Speech Recognition
vosk==0.3.45

# Environment & Config
python-dotenv==1.0.0

# HTTP Requests (for URL-based transcription)
requests==2.31.0

# WSGI Server (for local testing)
gunicorn==21.2.0

# Utilities
uuid==1.30
